feat(OAuth): refacto to make room for other authentication methods
